pyproject.toml
[project]
name = "slopepy"
version = "0.1.0"
description = "GPU-accelerated Digital Elevation Model (DEM) analysis tool for hillshade, slope analysis, geomorphology, and hydrology"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "cupy-cuda11x>=10.0.0",
    "numpy>=1.20.0",
    "rasterio>=1.2.0",
    "geopandas>=0.10.0",
    "shapely>=1.8.0",
    "scipy>=1.7.0",
    "typing-extensions>=4.0.0"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/slopepy"
"Bug Tracker" = "https://github.com/yourusername/slopepy/issues"
"Documentation" = "https://slopepy.readthedocs.io/"

[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "isort>=5.0",
    "mypy>=0.9"
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0"
]

[tool.setuptools]
packages = ["slopepy"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"